requirements.txt
# ======== CORE FASTAPI =========
fastapi==0.111.0
uvicorn[standard]==0.30.1

# ======== DATABASE =========
SQLAlchemy==2.0.31
alembic==1.13.2
psycopg[binary]==3.2.3


# ======== AUTHENTICATION / SECURITY =========
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9

# ======== ASYNC & WEBSOCKET =========
websockets==12.0
aiofiles==23.2.1

# ======== BACKGROUND TASKS / REDIS =========
redis==5.0.4
aioredis==2.0.1

# ======== VALIDATION & SERIALIZATION =========
pydantic==2.8.2
pydantic-settings==2.3.4

# ======== LOGGING & UTILS =========
python-dotenv==1.0.1
requests==2.32.3

# ======== OPTIONAL: DEV TOOLS =========
httpx==0.27.0
black==24.4.2
ruff==0.5.6
